LayoutTests/platform/mac/imported/w3c/web-platform-tests/mathml/presentation-markup/mrow/mrow-preferred-width-expected.txt


PASS Preferred width of mrow with mspace children
FAIL Preferred width of mrow with mn and mo children assert_approx_equals: mfenced expected 105.109375 +/- 1 but got 125.359375
blah
|
2
blah
|
2
blah
|
2
blah
|
2
